\name{scEMundir-package}
\alias{scEMundir-package}
\alias{scEMundir}
\docType{package}
\title{
\packageTitle{scEMundir}
}
\description{
\packageDescription{scEMundir}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scEMundir}
\packageIndices{scEMundir}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{scEMundir}

Maintainer: \packageMaintainer{scEMundir}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
